backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unmapped-backend"
version = "1.0.0"
description = "UNMAPPED - Skills-to-Opportunity infrastructure for the World Bank Youth Summit Hackathon"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "UNMAPPED Team" }]
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.0",
    "pydantic-settings==2.5.0",
    "httpx==0.27.0",
    "tenacity==9.0.0",
    "rapidfuzz==3.10.0",
    "pyyaml==6.0.2",
    "pandas==2.2.3",
    "structlog==24.4.0",
    "python-multipart==0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
]
embeddings = [
    "sentence-transformers==3.1.0",
]

[project.scripts]
unmapped-crawl = "scripts.crawl:main"
unmapped-seed = "scripts.seed_demo:main"
unmapped-build-taxonomy = "scripts.build_taxonomy:main"

[tool.setuptools]
packages = [
    "api",
    "api.routes",
    "config",
    "core",
    "crawlers",
    "matching",
    "scripts",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
